Fix: gen image create qcow in tmpdir if necessary
If the output directory is a filesystem that does not support lock
operations, qemu-img rebase will fail. We see this with lustrefs, though
it could happen with other filesystems.

Check for lustrefs and, if it's detected fallback to operating in /tmp.
After the file is produced in /tmp, copy it to the expected path.