Bugfix: Tests must create 64-bit binaries for 64-bit tests
We create a fake ELF to run penguin init on before we patch the config.
Previously we were creating 32-bit binaries and failing to check if the
identified architecture matched the expected one. It worked for aarch64
(surprisingly) but not for mips64eb: all prior mips64eb unit tests were
actually testing 32-bit mips little endian.

Tests are now failing as penguin init is incorrectly detecting the
mips64eb busybox as mips32eb.