fix(hyperfs): open /proc/self/fd/FD fail when FD is a pipe
Previously, hyperfs opened files inside /proc/self on behalf of the
calling process.  This led to problems when only the calling process is
able to access the files.  For example, you can't open the file
descriptors /proc/PID/fd/FD of another process when the file is a pipe.
This change makes /proc/self a symlink to /igloo/pfs/real/proc/self (the
real /proc/self managed directly by procfs) to have the process open the
files directly.  This means we can't define pseudofiles inside
/proc/PID, but we probably never will want to do that.