fix(hyperfile): do not shift hyperfs-buffer by offset
The offset is provided to indicate how far into a file we are
reading/writing. It is NOT related to the buffer that we're putting the
results into.

Previously we would calculate an invalid pointer once a FD had a large
offset and get stuck in an infinite loop where hyperfs pages in the
right data, we try to read the wrong data, fail to read, raise an error,
and repeat.